+++ _src/interfaces/serviceRequestInterceptor.ts
@@ -2,6 +2,7 @@ import { ExecutionContext } from './executionContext';
 
 export interface ServiceRequestInterceptor<TRequest> {
 	execute(
-		request: TRequest, context: ExecutionContext,
+		request: TRequest,
+		context: ExecutionContext,
 	): Promise<TRequest>;
 }